requirements.txt
matplotlib
requests
pollinations
pillow